requirements.txt
alembic==1.18.4
aniso8601==10.0.1
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
